Pulled packet parsing out into a function that you can use.
Just take all data in with recv as normal. Add characters (only as many as recv returns) to a string that acts as a buffer. Pass this string to parsePacket, and it will return a struct with a string and type. Use the type to identify which Google proto object to serialize the string with.